Ooga booga, trade cards are "working", I guess. But what is next? Well, trade is, ya know.
