+++ src/civilization/general/general_plugin.rs
@@ -14,7 +14,6 @@ use crate::{GameActivity, GameState};
 use bevy::app::{App, Plugin, Update};
 use bevy::prelude::{in_state, AppExtStates, IntoSystemConfigs, OnEnter};
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
-use crate::civilization::trade_cards::trade_card_plugin::TradeCardPlugin;
 
 pub struct CivilizationPlugin;
