+++ src/civilization/trade_cards/mod.rs
@@ -1,4 +1,5 @@
 pub mod trade_card_plugin;
 pub mod trade_card_components;
 pub mod trade_card_systems;
+mod trade_card_setup_systems;
