+++ src/civilization/trade_cards/trade_card_plugin.rs
@@ -3,7 +3,8 @@ use bevy::app::Startup;
 use bevy::prelude::{App, OnEnter, Plugin};
 use bevy_common_assets::ron::RonAssetPlugin;
 use crate::civilization::trade_cards::trade_card_components::CivilizationCardDefinitions;
-use crate::civilization::trade_cards::trade_card_systems::{acquire_trade_cards, load_civilization_cards, setup};
+use crate::civilization::trade_cards::trade_card_setup_systems::{load_civilization_cards, setup};
+use crate::civilization::trade_cards::trade_card_systems::acquire_trade_cards;
 
 pub struct TradeCardPlugin;
